add dataprocessing module in vmmath
