+++ pyLOM/vmmath/__init__.py
@@ -7,21 +7,23 @@
 # Last rev: 27/10/2021
 
 # Vector matrix routines
-from .maths      import transpose, vector_sum, vector_norm, vector_mean, matmul, matmulp, vecmat, argsort, eigen, polar, cholesky, vandermonde, conj, diag, inv, flip, vandermondeTime, time_delay_embedding
+from .maths          import transpose, vector_sum, vector_norm, vector_mean, matmul, matmulp, vecmat, argsort, eigen, polar, cholesky, vandermonde, conj, diag, inv, flip, vandermondeTime
 # Averaging routines
-from .averaging  import temporal_mean, subtract_mean
+from .averaging      import temporal_mean, subtract_mean
 # Truncation routines
-from .truncation import compute_truncation_residual, energy
+from .truncation     import compute_truncation_residual, energy
 # Statistics routines
-from .stats      import RMSE, MAE, r2, axiswise_r2, data_splitting
+from .stats          import RMSE, MAE, r2, axiswise_r2
 # SVD routines
-from .svd        import qr, svd, tsqr, randomized_qr, init_qr_streaming, update_qr_streaming, tsqr_svd, randomized_svd
+from .svd            import qr, svd, tsqr, randomized_qr, init_qr_streaming, update_qr_streaming, tsqr_svd, randomized_svd
 # FFT routines
-from .fft        import hammwin, fft
+from .fft            import hammwin, fft
 # Geometry and mesh routines
-from .geometric  import cellCenters, normals, euclidean_d
+from .geometric      import cellCenters, normals, euclidean_d
 # Regression routines
-from .regression import least_squares, ridge_regresion
+from .regression     import least_squares, ridge_regresion
+# Data processing module
+from .dataprocessing import data_splitting, time_delay_embedding
 
 
 del maths, averaging, truncation, stats, geometric, regression
